Updating plumbing and type checks for yii2 craft3
