+++ composer.json
@@ -3,10 +3,23 @@
     "description": "A helpful inspector plugin for Craft CMS",
     "homepage": "https://github.com/amacneil/craft-inspector",
     "license": "MIT",
+    "version": "1.0.0",
     "authors": [
         {
             "name": "Adrian Macneil"
         }
     ],
-    "type": "craft-plugin"
+    "autoload": {
+        "psr-4": {
+            "amacneil\\inspector\\": "src/"
+        }
+    },
+    "type": "craft-plugin",
+    "extra": {
+        "handle": "inspector",
+        "name": "Inspector",
+        "class": "amacneil\\inspector\\Plugin",
+        "developer": "Adrian Macneil",
+        "developerUrl": "http://adrianmacneil.com"
+    }
 }
